fixes issue 33900: Wrong decimal default value having ',' as decimal separator
When using a decimal value as default for a numeric column, this very same value is a String passed to the createFromClassicString of NumberUIDefinition which is used by the FIC to retrieve the column value. The default values are part of the code, so in case of decimal values, they must me defined as a valid Java decimal number. But the createFromClassicString expects a value formatted according to the Format.xml file.

For this reason, now the decimal default values are formatted properly. Although this value will be converted again in the createFromClassicString, this is the cleanest solution because for the rest of the flows within the FIC, the createFromClassicString method receives a value formatted according to the to the Format.xml file.