Fixed issue 33871: Check quantity precision (as per product UOM precision) when entering/editing ticket lines in POS
There is no check to verify the precision of a unit of measure.

With this solution, the precision of a unit of measure is check and, in
case the value has more precision than defined, an error is raised.

--HG--
extra : rebase_source : b7697d3290c029cfadb8537fa5f1bdbb8b261b32