Fixes issue 33813: Wrong payment date when voiding an unpaid invoice
If voidDate is not null (Purchase Invoice), voidDate, which is the reversed invoice's date, will be taken as dummy payment's date.
If voidDate is null (Sales Invoice), new Date(), which is the reversed invoice's date, will be taken as dummy payment's date instead of invoice.getInvoiceDate(), which was the original invoice's date.