Fixed issue 33799: AD_ROLE_ID was not working fine in the tabs display logic
The display logics using the AD_ROLE_ID session variable were not working fine. The problem was in the ob-standard-view.js class. The session attributes were overwritten with the values saved in the OB.Properties. And in this case it was null.
To fix this, a condition has been added, now if the OB.Properties(propertyName) is null, the value of the session attribute is not overwritten.

--HG--
extra : rebase_source : 3560e608e341f8683d6c98341e20e3273ce48fd4