Fixed issue 33861: Sequence number is getting incremented when empty bills are deleted through end session
When a logout is done, in the standard flow, all the tickets present in that
moment are deleted even if the receipt has no lines and is completely new.

This is a problem when the preference to Save Delete Tickets is active.

The solution, is to delete all the tickets with a document sequence lower than
the last ticket with lines.