Fixed issue 34015. Removed slow query getting Organization Ledger.
The method to get the default General Ledget was using a query that filtered
and sorted the results using ad_isorgincluded procedure. In case of having a
large number of organizations this query was slow.

The method is now using the OrganizationStructureProvider to get the parent
organization List of the required org to search the default General Ledger.

--HG--
extra : transplant_source : %DC%AFQ%95%23z%B2%7B%1D%8B%90%17%89Mb%5D%9Dl%AD%B6