[pi-accessibility-improvements] LeftColumn layout of the workspace are now accessible
Before there was not possible to access the left side members of the workspace; recent views, Recent Documents and Manage Workspace; using the keyboard. It was only possible to access them using the mouse. Now, if the "Enable full screen reader" preference is enabled, it is possible to access them using the keyboard.