[pi-accessibility-improvements] Problems with the toolbar item names fixed
There were some problems with some item names in the toolbar. Some of them where saying "Dot" at the beginning of the name of the button and the attachments and the saved views buttons were totally wrong.
To fix the first issue, the title: '.' of the class ob-toolbar.js has been changed for title:''. Now the word "dot" is no longer said.
To fix the second issue, the title for the attachments button and the save views button has been set.