[pi-accessibility-improvements] Fixed the aria label of attachments
Before it was set a title for the attachments toolbar button. But it was a static name, it was not correct because the name needs to change if there are attachments or if not. So, the method setAriaState(state, stateValue) is invoked to change the aria-label property in the different situations.