Fixes issue 34043: There is a restriction when extending a callout
There was a restriction when extending a callout from another callout and the same combo was updated.

This problem is fixed and now it is allows to delete any entry of a combo previously updated. Besides it
has been improve how to manage combo entries by using a map. It has been improve some comments and
variable names too.