Fixed issue 33983: Last payment in foreign currency is adjusted to expected to avoid differences
When a payment is done, the code will check if this payment satisfy the amount which is pending in foreign amount. If the amount is satisfied then the amount of this payment will be calculated from the pending amount instead of converting the payment to the default currency which can cause differences. Reuse value stored in var instead to call function again