Fixed issue 34004: [SERQA 1709] Not homogeneous keyboard use
The actual implementation of RFID doesn’t allow to separate it from the
scanMode variable and the Use external Input flows.

The solution is to force the stop of the RFID when the payment tab is showed.
Due to this change, it’s necessary to implement a new function to recover the
focus in the focuskeeper input.

--HG--
extra : rebase_source : 1a95977e061c3f3aa21a8ce96a036369f791f5fc