Fixes issue 33977: "Set new currency" process is converting credit properly
When the "Set new currency" process is configured to do not use a default rate and a manually rate is introduced, it is used for recalculate the generic credit instead the current conversion rate. Also, when more than one credit payment exists for the business partner, the process was generating the new entries with the last original value of the generated credit of the payment, and it caused that all generated credits were the same. Now, the generated credit is always recalculated with the original generated credit and rate of the process.