Fixes issue 34067: No UOM Conversion Rate between UOMs found warning shown.
This problem is related with Improve performance callouts project. The problem
was related with how to manage null and "null" values.

Previously it was necessary to send the "null" value as part of the callout
execution result because StringBuilder in CalloutInfo class couldn't accept
null Objects values. Then, this way the parseCalloutResponse() of the FIC
was able to retrieve the null values properly. This now is not necessary in
SimpleCallouts due to its new internal implementation.

It have been resolved this problem by sending null Object as part of the
callout execution result in both cases: When a "null" String or null
Object is recieved. Now the problem it has fixed and the null and "null"
values are managing properly.