Fixed issue 0034243: Generic products are not shown
Backed out changeset 10ca4126a517

It is not an issue, is a design defect. It is a limitation of the incremental refresh. Incremental refresh, will remove products from local database only when active column is unchecked. If the rest of columns of the product change, it will be loaded in the total refresh.
Not only sale column affects to the product to be loaded or not, also if we remove it from the assortment or from the pricelist...

--HG--
extra : rebase_source : 58107bfa6544d8ca8a5399c2171b9e114e5d827d