Fixed issue 34057: The focus' behavior was not the expected.
In some windows, in form view, the focus was not working fine when using the tab key. In some cases the focus was jumping from a field to the child tab bar instead to the following focusable field, and after, to the field.

The problem was that the tabIndex of the child tab bar was the same than the tabIndex of the field which was expected to be focused.
This is caused because the tabIndex of the child tab bar is not being updated when the formitems are created, and that is the reason because the tabIndexes are overlapped.

To fix this, the "prepareViewForm" of the "ob-standard-view" class has been modified, now it forces to update the child tab bar tabIndex and it gives to it a higher value than the values of the tabIndexes of the parent elements.

--HG--
extra : rebase_source : d49c7e3bf2a8c8c0dbd4d7f0aa7c12e533eb2810