Related to issue 33939: SER QA 1790 Wrong remaining to pay amount when paying with multicurrency
The financial account selected to use in cancel and replace has changed. Now the
query searches for  those financial account with the same currency as defined at
organization level.

The second change use the previous flow to generate the information of the
payments in case the order is a cancel and replace ticket. In the rest of the
cases, the information of the payments are read directly from the real payment
(in it’s original currency).

--HG--
extra : rebase_source : be6a09188ff0a81156dc7b6db7d9c3a1a98f7f53