Fixes issue 34323: TriggerHandler only removes the entries created by itself
The TriggerHandler class adds entries to AD_SESSION_STATUS with the isimporting column set to 'Y' to soft-disable the Openbravo triggers. Other processes may want to use the AD_SESSION_STATUS table to disable/reenable other features. The problem is that when TriggerHandler reenables the triggers, it removes all the contents of AD_SESSION_STATUS, not only the records created by the TriggerHandler itself.

To fix this, now TriggerHandler only removes from AD_SESSION_STATUS the rows whose isimporting column is 'Y'.