Related with issue 34323: Supports adding columns to AD_SESSION_STATUS
The TriggerHandler.disable methods inserts a row in AD_SESSION_STATUS to soft-disable the triggers. The problem is that the insert statement did not include the list of columns names in the INSERT statement, resulting in the statemeng failing if any module adds a new column to the AD_SESSION_STATUS table.

--HG--
extra : rebase_source : d53b01397e63b6b0c7245a1d865bc34de76c1058