Fixed issue 34061: The drop of the constraints is now more restrictive
The ant update database task has been modified to be more restrictive when droping the constraints and deleting rows. Before this project it was done  for every FK constrainst to AD tabes. And after that every of them where regenerated. This caused performance issues.
Now, it is only done when deletion or insertion changes are detected, and for those tables which are affected.