Fixes issue 34226: Requisition lines with same product are correctly grouped
Selection of lines in Requisition to Order process always try to maintain the
original order of the lines where the order is created from.
To do that, it first order by line number and then by product, this guarantees
the order but lines with same attributes set (product, attribute, description,
price and UOM) could be no consecutives and not correctly grouped.

To make a correct grouping, was changed the logic of always expect consecutive
products. Before creating the new purchase order line all existing lines
with same attribute set are grouped into a unique line with the total of the
order quantities in a Hash Map object. Then, purchase order lines are created
from the Hash Map containing lines with totals of the originals lines with same
attributes set.