Fixes issue 34332: Purchase Order field is empty in Purchase Invoice header
Purchase Order field was empty in Purchase Invoice header when selecting the Goods Receipt
in Create Lines From process, because the Purchase Order Id of Goods Receipt was not take into account to fill
Purchase Order field in Purchase Invoice header. Now the Purchase Order Id of Goods Receipt is retrieved to fill this field.