Fixed issue 33915: [SER QA1632] Cash Management synchronization is failing due to missing information related to cashup
With the new infrastructure, the update of the cashup is sent with the
information of the synchronization models. In this case, the problem was the
information sent to the backend about the payment methods of the cashup.

Now, the cashup information is compose before it is sent to the backend.

--HG--
extra : rebase_source : 5b97c99f6970a3f34931debcb63801593aee3c18