Fixed issue 34405: The Manage Variants filter was not working fine.
If the filter was used, after that a record was selected and after that the filter was cleaned up, the "no items to show" message was shown instead of showing all the records.
The problem was that the "ManageVariantsDS" manual datasource was not capable of managing the criteria: "criteria:{"fieldName":"id","operator":"notNull"}" which is the criteria that must return every record.
The problem was in this code line: "value = criteria.getString("value")"  because the creiteria has not value, it only has fieldName and operator. It was throwing an exception which was not being correctly handled in the catch. This problem has been solved also, by properly handling the exception by showing log.
The first problem has been solved by processing the concrete criteria properly. If this criteria is found, it means that every record must be shown, so it continues with doing nothing.
There was a another problem that has been solved and it is that the searchKey and the nameFilters were acting as a "Contains" filter instead of like as an "iContains".
This has been fixed by changing the "contains" method of StringUtils by the "containsIgnoreCase".

--HG--
extra : rebase_source : 0a38bfb4e6c08a8168eb35c64e69118934571511