Fixes issue 34410: GL item in Open Items section is set read only
If Simple G/L Journal is multi general ledger, the G/L Item in Open Item section
is set read only and it's value is set to G/L Item in the line, this value is
updated when G/L Item value in the line changes.

The display logic of G/L Item in Open Item section takes into account if Multi
General Ledger is set in header. A new callout was made to kep updated it's value
when Account values in line changes.