Fixes issue 26499: Removed @FIN_NoMultiAccountingAllowed@ restriction for G/L
When the Simple GL Journal window was implemented it included to be able to
"create G/L Item payments from journal" for organization with several General
Ledgers. As this feature is available, only pending point was to remove
@FIN_NoMultiAccountingAllowed@ restriction (FIN_AddPaymentFromJournalLine.java)
when processing a document defined as "Multi-General Ledger", and it is done by
this patch.