Fixed issue 33850: The Location/Address filter was not working fine
The location/Address filter of the Location/Address subtab of the "Business Partner" tab was not working fine when trying to filter by Location/Addresses which had not a Region defined.
The problem was in the generated query, there was a concatenation with the region name, and when the region was null, all the result was null.
To fix this problem the coalesce funtion has been added to the query in the cases where it was not added. Now, if the value is null, as it is the case, the empty string will be get in the concatenation.

--HG--
extra : rebase_source : 8b76ce623da22cda27cb49c12f9d36d2f2164a4f