Fixed issue 34469: The manage variants record selection was not working fine.
When a record was selected and after that another action was done, the record which must remain selected, was not.
The problem was in the "ManageVariantsDS". There was a property which was not being set. In the client there is a place where it is searching the selected records by a "findByKey" method. This method was searching the property "primaryKeys" and it needs to have the id set. In this case, as it was not set, it was not returning anything and it was not being able to find the selected record. That was causing the problem of the unselected record.
The fix consists on implementing the "getDataSourceProperties" method and there set the needed properties of the datasource.

--HG--
extra : rebase_source : d0daca3fb8b55fc3281e76cf52804afb30505911