related to issue 34499: removed PreferenceEventHandler class
PreferenceEventHandler class was being used to detect changes on the preference values cached by the CachedPreference class. As CachedPreference now declared the @SessionScoped is no longer a singleton therefore the cached values will persist along the life cycle of a session.

As preferences are refreshed with every new session this handler is no longer required. Besides, the injection of the @SessionScoped bean in this handler was causing WELD-001303 error.