Fixes issue 34286: Product selector allows wrong storage bin in shipment window
The Product Complete selector, when used in goods shipment window, allowed select
products in storage bin not belonging to the on-hand warehouses.

A new column was added to the M_PRODUCT_STOCK_V view to retrieve the warehouse
and the filter expresion of Product Complete selector was modified to include
the filter by on-hand warehouses, when used in goods shipment window.