fixes issue 34485: Record marked as changed after printing in grid view
After printing, just the selected record is refreshed with the refreshCurrentRecord function. The problem was caused because that function was retrieving all the properties instead of just the required grid properties. Eventually, this was causing that some extra fields where being compared when opening the form (in recordHasActualChanges function) causing a wrong detection of changes.

To solve the problem now, when refreshCurrentRecord function is executed in grid view, it retrieves just the grid selected properties. Note that this is the behavior when refreshing the whole grid.