Fixes issue 34612: Performance problem in Return To Vendor Shipment
Implemented C_OrderLine.QtyDelivered for return to vendor flow. This column was used only by the sales flow. However RTV (which is just a return in a purchase flow) can be considered like a sales flow (actually we are delivering products), so it has been easily adapted.
This column is now shown in RTV lines.

Implemented C_Order.IsDelivered flag for sales orders and return to vendor flows. This column was only implemented for the following Sales orders subtypes: "On Credit Order", "POS Order" and "Warehouse Order"; now it should work fine for all the sales order (although this is not a requirement for the RTVS refactor, it has been quite easy to do it).
This column is now shown in RTV header.