[pi-process-multi-record] processes executed for more than one record
Now, it is possible to select more than one record and execute a process. Before this fix, even it was possible to select more than one record, the process was only affecting the first selected record. The javascript logic which handles this has been added.
Also, a refresh of the view, in the cases of having multi record processes, has been added.