Fixed issue 33792:In P&E windows, filter drop down options disappear sometimes
The problem was in P&E windows using filter dropdown. First time all were working fine but the second time, when filter dropdown was open and a filter option was previously selected, filter dropdown options were disappear. To reproduce this problem it was necessary to have a selected record in P&E.

When filter dropdown options were disappear, the request had some extra parameters. This extra parameters were a problem in order to build the query properly. These were the extra parameters:
	_constructor:AdvancedCriteria
	_OrExpression:true
	operator:or

At the end these parameters build a bad query because main alias was not managed properly in the server side. As a result of the operator parameter (operator:or) main alias was not use properly. The query was built poorly due to the aliases.

To solve the problem now, when cleanCriteria function is executed, it is also executed cleanOrCriterion function in order to clean criteria properly.