Fixed issue 34051: The tab key was not focusing properly in some circumstances
The fields were not correctly focused when using the tab key. If there was an image in the form, the focus was jumping to it when it was supposed not to.
The problems seems to be in the canvas item. The tab index was not being properly set.
To fix this the "ob-formitem-image.js" has been modifyed, now, the "canFocus" property has been set to true for the image canvas.

--HG--
extra : rebase_source : 1f1f91bb4181cc30c8b2fd4e7988bedf0c396cc6