Fixed issue 34324:Summary functions do not work in tabs based on HQL queries.
The summary functions didn't work in tabs based on HQL queries because this
functionality was not developed. The HQLDataSourceService hadn't the code to
managed summary functions properly.

To avoid this problem, it has been resolved by blocking the use of the summary
functions in the tabs based on HQL query.