fixes issue 34462: Process Definition menu icon should depend on the UI Pattern
The report icon is now selected both when using the quick launch search and the application menu. For this second case, now the UI Pattern is also sent to the client as part of the process definition menu entry. This helps to identify the type of process definition when retrieving the icon for the application menu entry.