Fixes issue 34661: You can set 2 conversion rate in the same period
In conversion rates window, it was possible to define different conversion rates
to the same currency in the same period. When inserting or updating a conversion
rate, the date was validated so no overlaping ocurred, but it was only validating
the cases when dateFrom or dateTo were between the newly created dates, not when
those dates were outside the range of dates, i.e, were overlaping the entire range
of dates already in the system for the given date.