Fixed issue 33606: The Preferences.getPreferenceFromList improved.
A continue has been added to do an early return in a for loop of the Preferences.getPreferenceFromList method. Putting this code there prevents a lot of executions needed to get the "winVisibilityNotDefined" and the "sameWinVisibility" variables.
The "isCurrentPreference" variable needs to be true to make this piece of code true: "if (isCurrentPreference && (winVisbilityNotDefined || sameWinVisibility)) {" and to return the listPref value. So, if the "isCurrentPreference" is false, it doesn't matter the values of the "winVisibilityNotDefined" nor "sameWinVisibility" values. That is why is a good idea to put the early return. In this way a lot of executions to get the variables will be avoided.

--HG--
extra : rebase_source : eb934d981939ce643846d5922fcff166444d14fc