Fixed issue 34521.  [Performance] Discounts calculation is slower when the user increase the number of lines of the receipt
The problem with this issue is that the functionality of findUsingCache was
never use, because the parameter to stablish the initial date to take a
promotion was always changing with the value of the actual date witch was
a datetime value.

In the backend, it’s not possible to define a promotion neither the starting
date nor ending date fields with a date including hours, minutes and seconds.

The solution is to parse the date to sue only dates without time for the query
criteria.