Fixes issue 34700: Warehouse name cut when Valued Stock Report exported to Excel
Warehouse name was cut when Valued Stock Report is exported to Excel. The problem
was that the field storing the warehouse name had a small width, so when the report
is exported, it cuts the name if it exceded the field's width. Since the warehouse's
name is stored in a varchar(60) in DB, the field's width was expanded to cover
the full length of the name.
Note: this problem can be extended to the remaining fields of the report.