Fixes issue 34699: Warehouse combo shows no values in Valued Stock Report
Before report was launched for first time, warehouse combo was not showning proper
values, it was empty. It was caused because this combo didn't get updated in the
onChange event of the Organization's combo, but it was updated when reloading the
report or running it with an organization already selected. Also, when selecting an
organization with no warehouses, but with children organizations that has a warehouses
asociated, the combo was not showing those warehouses. This has been fixed so when
selecting the organization with no warehouses, it now searches in its children in
order to show the warehouses if they exists.

--HG--
extra : rebase_source : eeceabe771e80da79fa2b035046a83827edf56b9