Fixed issue 33853: There were too much DS request
When a column with a summary function was set, if any of the grid columns was reordered, there were three DS requests triggered. To fix the problem two different things have been done. The first, in ob-view-grid.js, consists on setting the showGridSummary property to false before the call to the function "reorderField" of the Super- This avoids one flow of smartclient which was, at the end, triggering two DS requests.
The second approach is also in ob-view-grid, but also in ob-grid. The problem was that when the ob-grid.getSummaryRow function was invoked, the return statement was calling the "getSummaryRow" function of the Super. There, once again a DS request was triggered. To fix this, a new object has been created. This object contains three properties. One of them is a property which is true if the columns have been reordered.
Now, in the ob-grid.getSummaryRow, if that property is true and if there is a summary function, the summary row is returned.

--HG--
extra : rebase_source : 2da9dd54d97ceecb5276f12e71b63e83044f806a