fixes issue 34716: sometimes action buttons are displayed when they should not
After closing a new record in form view without saving the changes (either using the cancel button of the toolbar or the close button of the form) some action buttons were being displayed when they should not. This was happening because the form was not being restored as not new after closing it without saving the changes. Back in grid view, having the isNew flag as true was avoiding the required call to the FIC in SETSESSION mode which is needed to recalculate the display logic of the buttons.

--HG--
extra : rebase_source : 4481a6f75c7f45578c45c56c92622de296d7fb62