Fixes issue 34713: getProductCost() is not checking if product is production
If a product is defined as "production", when costing is checked for
that product, "getProductCost" method in AverageAlgorithm.java was filtering by
"warehouse" dimension (in case costing rule is defined like this) so error "No
average cost found for given product and date" was raised, when it should not.
The proposed solution has been implemented, adding a validation when constructing
the "where" condition, so now there is no error.