Fixes issue 32902: SelectAll checkbox is not currently active in all flows
The problem was that SelectAll checkbox was not currently active in P&E windows when all
lines were selected in grid.

The problem has been resolved by using updateCheckboxHeaderState() function. This function
update the state of the SelectAll checkbox when a P&E window have selected all of their lines.