Fixed issue 34677:Tooltip message of checkbox in P&E should be displayed.
The problem was that the tooltip message of the checkboxes in P&E should
be displayed. The messages should be "Click to (un)select all records"
and "Click to (un)select this record" depends on the checkbox.

The problem has been resolved by overriding setFieldProperties() method
in order to support hover for the checkbox field.