Fixed issue 33626: Unneeded ds request when cancelling a record fixed
There was an uneeded ds request when there was a summary function and a new record edition was cancelled in grid view.
To fix this, the showGridSummary property has been set to false before the call to the function "handleUpdate". This avoids one flow of smartclient which was triggerin a ds request.
After that the "setShowGridsummary" function is invoked, to show again the summary function values.

--HG--
extra : rebase_source : 06086076389012f46418520c6a4d728ea965b3d8