Fixes issue 34731: Wrong amount value when adding payment to a transaction
When creating a withdrawal transaction in Financial Account screen, if you add a
payment, the field Actual Payment was showing 0.00 instead of the correct amount
of the transaction. This issue is solved by calling the method convertToSafeDecimalString,
which returns a well formed string, so the correct amount can be displayed.