Related to issue 33853: Code review improvements
In js variables must be declared at the top of the function.
Also there were problems when reordering if there was not a summary function declared. This has been prevented with an if statement. A new function has been created, this, handles the behavior of the summary functions' grid when it is necessary.

--HG--
extra : rebase_source : 1e794045109de4a04f77a1cdd6d7d946f83d4a87