Fixed issue 33626: Code review improvements
When a record was cancelled i grid view, without any summary row defined in any column, the summary function grid was shown with the text "no items to show" this has been fixed by adding a condition to the ob-view-grid.discardEdits function. Now, the summary grid will only be shown when there actually exists a summary function in any column.