Fixed issue 34523: Two OBScheduler fields are now private
There was a problem with the "dateTimeFormat" and the "sqlDateTimeFormat" fields of the OBScheduler class. These static fields were public, so, anyone could access them and change their value. That could cause problems, for example in clustered envirnoments.
The OBScheduler is a singleton, so it makes no sense to have those fields set as static.
Now the fields are now declared as private, so the proper getters have been created.

--HG--
extra : rebase_source : 50d6d5d2dbac89e098a6206675474812eb177bf1