[PrePayments] Applied UI feedback.
The delivery button is changed from 'Can be delivered' and 'Cannot be delivered' to 'Set as Deliverable' and 'Set as Undeliverable'.
When selecting lines with different deliverable value, instead of hide, the value is shown as '<Multiple Values>'.
When selecting a line and a service line, the deliverable value is set to '<Multiple Values>'.
The 'Can be delivered' property is changed to 'Deliverable'.

--HG--
extra : transplant_source : %1B%0C%86%21%0B%D6p%DD%1B%C5E%C3%02%F5%984%159%B7%C3