Fixed issue 34830: Product selector of the Requisition subtab appears empty
The problem was introduced in a refactor of checkFetchDatasourceAccess() method of
BaseDataSourceService class. This refactor change the way in wich entities of
selectors were calculated. It is important to note that the refactor is not bad in
itself. The problem was appeared because it was forgotten to do an additional thing.

Before this refactor in some cases entities were not calculated properly. For example
in the subtab of Requisition window there are a complex selector of product. (Reference:
Product Complete - Parent Reference: Search). In this example, checkFetchDatasourceAccess
method was checked Product entity instead of ProductStockView. The refactor fixed this
problem.

Otherwise, the refactor missed to take into account entities of complex defined selectors
as derivedReadableEntities. To solve the problem now, when initialize() method of
EntityAccessChecker class calculates permissions, it will take into account this entities.