Fixed issue 34863: message while logging in is not properly translated
While executing the login and checking if the users password is expired or not (DefaultAuthenticationManager.java), the session variable of the language is empty so it always uses English language by default. For this reason 'CPExpirationPassword' message was not properly translated.

In order to reach an homogeneous behaviour related with translations it has been added language in LoginHandler class. Now the problem has been fixed and message is translated properly.