[multi-pool] Use putIfAbsent to avoid possible thread unsafeties
The putIfAbsent method of ConcurrentHashMap associates a key with the given value if the key is not already present in the map. This method is performed atomically, this way we ensure that two threads executing the getInstance(String pool) method at the same time always return the same instance for the same pool name.