Fixed issue 34864: The GL item section of add payment was not working fine
The grid of the G/L item section of the Add payment process was not workinf fine. The problem was that it was shown as sortable and as filterable and it must not be filterable nor sortable. This has been caused by the changeset[1]. This proyect set every column as sortable and as filterable and so was the problem.
To fix this the values of "canSort" and "canFilter" have been changed in the needed cases, which are the ones that satisfy the condition of having the canDelete property or the canAdd property set to yes and are in Pick and Execute window.
The code which handles the logic has been moved from the OBViewFieldHandler.getGridFieldProperties() to UIDefinition.getGridFieldProperties().

[1]http://code.openbravo.com/erp/devel/pi/rev/afa165be3e5900d1114d33e5106114d0e089876a

--HG--
extra : rebase_source : 425f1d029182631fa93ba5afcdb127803231ba4d