[multi-pool] SessionHandler must take into account the availability of the pools
As the SessionHandler is not removed anymore after doing commit or rollback, it remains present along all the lifecycle of the thread (it is removed at the end). But the session of a particular pool is indeed closed after doing commit or rollback, therefore the isSessionHandlerPresent() needed to be updated to take into account for which pools in particular is the session handler available.

From now the SessionHandler maintains a set with the names of the pools where a commit or a rollback has been done. This set is consulted to decide if the session handler is present for a particular pool as 'sessionHandler.get() != null' is not enough to decide it, because of the reason explained above.

Without this changeset, two consecutive calls to OBDal.getInstance().commitAndClose() would fail because the second commit would be performed without an active session and the checkInvariant method of the SessionHandler would throw an exception.