Fixed issue 34786: More Payment Methods popup has been refactored and it works fine after complete a sale
The popup is now defined in the posterminal window (As we do with the majority of the popups). Buttons which are in the list are generated dynamically when the popup is opened for the first time or when new payments methods have to be added.

--HG--
extra : rebase_source : a3775abcf1b3970d682acfcc946c53200a900199