Fixed issue 34950: A null check condition added
There was the following condition in ob-parameter-window-view.js "if(this.button.multiProcess)...". The problem here is that there is no a null check for the "this.button" and in the cases of opening other pick and execute window inside another one was failing because the view could not be refreshed.