Related to issue 34345: Code review improvements
Some code review improvements to check the amounts are GREATER THAN (not equals)
and/or LESS THAN (not equals) the filters provided in the UI. If there is not any
amount filter defined then it is not necessary to include a "HAVING" clause to the
query.

Applied these changes to Sales and Purchase Orders reports.