Fixed issue 34966: Filter by stock in product search does not allow to filter by other filter categories using part of the name of the product
In general, the search of products without remote is done using the contains
condition because we can assume the quantity of products will be small. This is
not done in case we use high volumes, where we search products using the
startswith condition.

With stock criteria the search is forced to do in the server. This filtering was
failing with the search of characteristic values and brand values.

--HG--
extra : rebase_source : ed4cc073393cf4e9e7236c1df73b2b7a0de91af6