Fixed issue 35020: If you have a payment method Cash without Count Cash fails on login
The solution is to execute sequentially the process of the creation of the
cashUp.

Now the execution of the hooks is done after the cashup is created and the
current cash is calculated.

Remove a callback parameter from initCashUp that overrides the original
callback.

--HG--
extra : rebase_source : d83f495259effa8bea4338e95f093199fbe5ed31