Fixed issue 35032: Preference “Web POS window Cash Up Partial” using a master or slave terminal does not work correctly
Finally, the preference is not the cause of this problem.

The problem was a function which waits until data is send to the backend. This function is triggered also when, in the WebPOS, we use the busyUntilFinishes functionality. In this case, the function reaches its timeout and then the result is shown.

The solution is to manage correctly the end of the function.

--HG--
extra : rebase_source : 8527086d3d244963b79afc039281d3d449c1eeb8